refactor: 불필요한 커스텀 예외 클래스 삭제 및 패키지 이동
