+++ src/main/java/roomescape/config/CheckAdminInterceptor.java
@@ -4,8 +4,8 @@ import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.web.servlet.HandlerInterceptor;
-import roomescape.controller.AuthenticationException;
-import roomescape.controller.AuthorizationException;
+import roomescape.controller.exception.AuthenticationException;
+import roomescape.controller.exception.AuthorizationException;
 import roomescape.domain.Member;
 import roomescape.domain.Role;
 import roomescape.service.MemberService;
