+++ src/main/java/roomescape/controller/dto/TokenResponse.java
@@ -1,4 +1,4 @@
-package roomescape.service;
+package roomescape.controller.dto;
 
 public record TokenResponse(String assessToken) {
 }
