+++ src/main/java/roomescape/controller/exception/AuthenticationException.java
@@ -1,7 +1,7 @@
-package roomescape.controller;
+package roomescape.controller.exception;
 
 import roomescape.exception.RoomescapeException;
-//TODO 패키지 분리
+
 public class AuthenticationException extends RoomescapeException {
 
     public AuthenticationException(final String message) {
