--- src/main/java/roomescape/domain/exception/InvalidDateException.java
@@ -1,10 +1,0 @@
-package roomescape.domain.exception;
-
-import roomescape.exception.RoomescapeException;
-
-public class InvalidDateException extends RoomescapeException {
-
-    public InvalidDateException(final String message) {
-        super(message);
-    }
-}
