--- src/main/java/roomescape/domain/exception/InvalidNameException.java
@@ -1,10 +1,0 @@
-package roomescape.domain.exception;
-
-import roomescape.exception.RoomescapeException;
-
-public class InvalidNameException extends RoomescapeException {
-
-    public InvalidNameException(final String message) {
-        super(message);
-    }
-}
