--- src/main/java/roomescape/domain/exception/InvalidTimeException.java
@@ -1,10 +1,0 @@
-package roomescape.domain.exception;
-
-import roomescape.exception.RoomescapeException;
-
-public class InvalidTimeException extends RoomescapeException {
-
-    public InvalidTimeException(final String message) {
-        super(message);
-    }
-}
